Replace allegedly all `width`, `height` and `rows_used` types with `usize`

+++ extensions/native/circuit/src/branch_eq/cuda.rs
@@ -43,8 +43,8 @@ impl Chip<DenseRecordArena, GpuBackend> for NativeBranchEqChipGpu {
         unsafe {
             native_branch_eq_cuda::tracegen(
                 trace.buffer(),
-                padded_height as u32,
-                trace_width as u32,
+                padded_height,
+                trace_width,
                 &d_records,
                 &self.range_checker.count,
                 self.timestamp_max_bits as u32,
